marker, libxlator: use long values for time management
Consistently use 'unsigned long' to manage time values, add convenient
'gf_net_time_to_host_time()' and 'gf_host_time_to_net_time()' to control
byte order where needed, adjust related code.

Updates: #1000

+++ xlators/features/marker/src/marker.h
@@ -84,7 +84,7 @@ enum {
     } while (0)
 
 struct marker_local {
-    uint32_t timebuf[2];
+    unsigned long timebuf[2];
     pid_t pid;
     loc_t loc;
     loc_t parent_loc;
